webllisto_backend/requirements.txt
Babel==2.6.0
Django==2.2.1
django-phonenumber-field==2.4.0
djangorestframework==3.9.4
phonenumbers==8.10.12
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
